move MetcalfScoring to independent file
